modif url item

+++ app/Controllers/ItemController.php
@@ -10,18 +10,6 @@ use Psr\Http\Message\ResponseInterface;
 
 class ItemController extends MainController
 {
-    public function getItemCreate(RequestInterface $request, ResponseInterface $response)
-    {
-        $get = explode("=", $request->getUri()->getQuery());
-        $liste = Liste::where('token', $get[1]);
-        if ($liste->count() == 1) {
-            $liste = $liste->first();
-            $this->render($response, 'pages/itemCreate.twig', ["current_page" => "itemCreate", "liste" => $liste]);
-        } else {
-            $this->render($response, 'pages/404.twig', ["current_page" => "404"]);
-        }
-    }
-
     public function getItems(RequestInterface $request, ResponseInterface $response, $args)
     {
 
@@ -62,56 +50,53 @@ class ItemController extends MainController
         }
     }
 
-    public function getItemManage(RequestInterface $request, ResponseInterface $response)
+    public function getItemManage(RequestInterface $request, ResponseInterface $response, $args)
     {
-        $get = explode("=", $request->getUri()->getQuery());
-        $item = null;
-        if ($get[0] != "") {
-            $item = Item::where('id', $get[1]);
-            if ($item->count() == 1) {
-                $item = $item->first();
-                $reservation = Reservation::where('id_item', $get[1]);
-                if ($reservation->count() == 1) {
-                    $this->render($response, 'pages/reserved.twig', ["current_page" => "reserved"]);
-                }
-                else {
-                    $this->render($response, 'pages/itemManage.twig', ["current_page" => "itemManage", "item" => $item]);
-                }
-            }
-            else {
-                $liste = Liste::where('token', $get[1]);
-                if ($liste->count() == 1) {
-                    $this->render($response, 'pages/itemManage.twig', ["current_page" => "itemManage", "item" => $item]);
-                }else {
+        if (array_key_exists('token', $args)) {
+            $item = null;
+            if (array_key_exists('id', $args)) {
+                $item = Item::where('id', $args['id']);
+                if ($item->count() == 1) {
+                    $item = $item->first();
+                    $reservation = Reservation::where('id_item', $args['id']);
+                    if ($reservation->count() == 1) {
+                        $this->render($response, 'pages/reserved.twig', ["current_page" => "reserved"]);
+                    } else {
+                        $this->render($response, 'pages/itemManage.twig', ["current_page" => "itemManage", "item" => $item]);
+                    }
+                } else {
                     $this->render($response, 'pages/404.twig', ["current_page" => "404"]);
                 }
+            } else {
+                $this->render($response, 'pages/itemManage.twig', ["current_page" => "itemManage", "item" => $item]);
             }
+        } else {
+            $this->render($response, 'pages/404.twig', ["current_page" => "404"]);
         }
     }
 
-    public function postItem(RequestInterface $request, ResponseInterface $response)
+    public function postItem(RequestInterface $request, ResponseInterface $response, $args)
     {
-        $get = explode("=", $request->getUri()->getQuery());
-        if ($get[0] == "item") {
-            if (isset($get[1])) {
-                $img = strip_tags($_POST['img']);
-                if (!empty($_POST['img']) && !str_contains(strip_tags($_POST['img']), "assets/img/")) {
-                    $headers = @get_headers(strip_tags($_POST['img']));
-                    if ($headers == false) {
-                        $img = "assets/img/" . strip_tags($_POST['img']);
+        if (array_key_exists('token', $args)) {
+            $liste = Liste::where('token', $args['token']);
+            if ($liste->count() == 1) {
+                $liste = $liste->first();
+                $item = null;
+                if (array_key_exists('id', $args)) {
+                    $img = strip_tags($_POST['img']);
+                    if (!empty($_POST['img']) && !str_contains(strip_tags($_POST['img']), "/mywishlist/public/assets/img/")) {
+                        $headers = get_headers($_POST['img']);
+                        if (!$headers || strpos($headers[0], '404')) {
+                            $img = "/mywishlist/public/assets/img/" . strip_tags($_POST['img']);
+                        }
                     }
-                }
-                $item = Item::where('id', $get[1])
-                    ->update(['nom' => strip_tags($_POST['nom']),
-                        'descr' => strip_tags($_POST['description']),
-                        'url' => strip_tags($_POST['url']),
-                        'img' => $img,
-                        'tarif' => strip_tags($_POST['tarif'])]);
-            }
-        } else {
-            if (isset($get[1])) {
-                $liste = Liste::where('token', $get[1])->first();
-                if (isset($liste)) {
+                    $item = Item::where('id', $args['id'])
+                        ->update(['nom' => strip_tags($_POST['nom']),
+                            'descr' => strip_tags($_POST['description']),
+                            'url' => strip_tags($_POST['url']),
+                            'img' => $img,
+                            'tarif' => strip_tags($_POST['tarif'])]);
+                } else {
                     $item = new Item();
                     $obj = json_decode($liste);
                     $item->liste_id = $obj->no;
@@ -120,11 +105,10 @@ class ItemController extends MainController
                     $item->url = strip_tags($_POST['url']);
                     $item->tarif = strip_tags($_POST['tarif']);
                     $img = strip_tags($_POST['img']);
-                    if (!empty($_POST['img']) && !str_contains(strip_tags($_POST['img']), "assets/img/")) {
-                        $headers = @get_headers($_POST['img']);
-                        echo var_dump($headers);
-                        if ($headers == false) {
-                            $img = "assets/img/" . strip_tags($_POST['img']);
+                    if (!empty($_POST['img']) && !str_contains(strip_tags($_POST['img']), "/mywishlist/public/assets/img/")) {
+                        $headers = get_headers($_POST['img']);
+                        if (!$headers || strpos($headers[0], '404')) {
+                            $img = "/mywishlist/public/assets/img/" . strip_tags($_POST['img']);
                         }
                     }
                     $item->img = $img;
@@ -132,22 +116,26 @@ class ItemController extends MainController
                 }
             }
         }
-        //$this->redirect($response, 'home');
+//$this->redirect($response, 'home');
         $this->render($response, 'pages/home.twig', ["current_page" => "home"]);
     }
 
-    public function deleteItem(RequestInterface $request, ResponseInterface $response)
+    public
+    function deleteItem(RequestInterface $request, ResponseInterface $response, $args)
     {
-        $get = explode("=", $request->getUri()->getQuery());
-        $item = Item::where('id', $get[1]);
-        if ($item->count() == 1) {
-            $item = $item->first();
-            $reservation = Reservation::where('id_item', $get[1]);
-            if ($reservation->count() == 1) {
-                $this->render($response, 'pages/reserved.twig', ["current_page" => "reserved"]);
+        if (array_key_exists('id', $args)) {
+            $item = Item::where('id', $args['id']);
+            if ($item->count() == 1) {
+                $item = $item->first();
+                $reservation = Reservation::where('id_item', $args['id']);
+                if ($reservation->count() == 1) {
+                    $this->render($response, 'pages/reserved.twig', ["current_page" => "reserved"]);
+                } else {
+                    $item->delete();
+                    $this->render($response, 'pages/home.twig', ["current_page" => "home"]);
+                }
             } else {
-                $item->delete();
-                $this->render($response, 'pages/home.twig', ["current_page" => "home"]);
+                $this->render($response, 'pages/404.twig', ["current_page" => "404"]);
             }
         } else {
             $this->render($response, 'pages/404.twig', ["current_page" => "404"]);
